Don't use std::sort on std::pair with a pointer component
I've gone through all sort calls in the DLL after reading this old commit message by karadoc:
https://github.com/f1rpo/AdvCiv/commit/300532663d710af4d2bd1ea5fde666f9eba214a7

Found just one potential problem, introduced by change [advc.108b]. Easy enough to use a (logical) id instead of a pointer.

+++ CvGameCoreDLL/WarAndPeaceCache.cpp
@@ -375,13 +375,13 @@ double WarAndPeaceCache::goldPerProdSites() {
 
 	PROFILE_FUNC();
 	CvPlayerAI const& owner = GET_PLAYER(ownerId);
-	std::vector<double> foundVals;
+	std::vector<int> foundVals;
 	for(int i = 0; i < owner.AI_getNumCitySites(); i++) {
 		CvPlot* site = owner.AI_getCitySite(i);
 		if(site != NULL)
 			foundVals.push_back(site->getFoundValue(owner.getID()));
 	}
-	std::sort(foundVals.begin(), foundVals.end(), std::greater<double>());
+	std::sort(foundVals.begin(), foundVals.end(), std::greater<int>());
 	int const sitesMax = 5;
 	double weights[sitesMax] = { 1, 0.8, 0.6, 0.4, 0.2 };
 	double sites = 0;
